fix(mcp-market): scope public registry access
